Adjusting week2 for more unit test instructions.
