+++ week2/week2_README.md
@@ -12,9 +12,15 @@ This week we will cover the following:
 
 # Week 2 Homework and Lab
 * Watch the content from the 'Week 2 Content' section below 
-* Complete the Week 2 Lab
+* Complete the Week 2 Lab at [week2_lab.py](week2_lab.py)
 
 ## Week 2 Content
  Use this content to learn about the topics and help with the week 2 labs.
  - [Instruction and videos](https://where-ever.com)
 
+## Week 2 Stretch Challenge
+In the week 2 stretch challenge you will implement a unit tested method body that requires Spark.
+We complete this exercise as a gentle introduction to unit testing.  If you want to know why unit
+testing is important, [please watch this video](https://www.youtube.com/watch?v=Ezz5nAH1opA).
+In later weeks we will do a deeper dive to disect the components of the unit test so by the end
+of this course you know how to write your own unit tests from scratch.
